+++ CineFlick/Movies/Models/JSON/Movie/MovieJson.swift
@@ -20,7 +20,7 @@ struct Results: Decodable {
     let vote_count: Int?
     let id: Int?
     let video: Bool?
-    let vote_average: Float?
+    let vote_average: Double?
     let title: String?
     let popularity: Double?
     let poster_path: String?
@@ -36,7 +36,7 @@ struct Results: Decodable {
         vote_count = json["vote_count"] as? Int ?? 0
         id = json["id"] as? Int ?? 0
         video = json["video"] as? Bool ?? false
-        vote_average = json["vote_average"] as? Float ?? 0.0
+        vote_average = json["vote_average"] as? Double ?? 0
         title = json["title"] as? String ?? ""
         popularity = json["popularity"] as? Double ?? 0
         poster_path = json["poster_path"] as? String ?? ""
